requirements.txt
Flask==2.0.1
mysqlclient==2.2.1
Pillow==10.0.0
pycurl==7.45.2
pyls-spyder==0.4.0
PyMySQL==1.0.2
PyQt5==5.15.10
PyQtWebEngine==5.15.6
setuptools==69.5.1
